+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-## [0.0.4] - 2019-05-27
+## unreleased
+
+## 0.0.5 - 2019-05-27
+
+### Changed
+
+- update package ts-loader
+
+### Added
+
+- Should now also fetch scoped packages
+
+## 0.0.4 - 2019-05-27
 
 ### Fixed
 
@@ -8,16 +20,16 @@
 
 - Escape hatch for issue [#1](https://github.com/JCofman/Changelog/issues/1) to avoid unnecessary fetching in files other then `package.json`
 
-## [0.0.3] - 2019-05-17
+## 0.0.3 - 2019-05-17
 
 - Tries to fetch GitHub releases if no changelog was found
 
-## [0.0.2] - 2019-05-16
+## 0.0.2 - 2019-05-16
 
 ### Added
 
 - New icon
 
-## [0.0.1] - 2019-05-15
+## 0.0.1 - 2019-05-15
 
 - Initial release